dont' override SERVER_* by the Host header
Gunicorn shouln't override the SERVER_* by the Host header. The client
will take care about it during the URL reconstruction if needed.

Since the spec don't support unix sockets, Gunicorn is now returning
(SocketPath, '') when using a unix socket which may or may not be a
problem with the current clients.